transcoder/requirements.txt
kafka-python
boto3
pydantic_settings
python-dotenv
requests